+++ inst/include/target/glm.hpp
@@ -1,7 +1,7 @@
 /*!
   @file glm.hpp
   @author Klaus K. Holst
-  @copyright 2018-2020, Klaus Kähler Holst
+  @copyright 2018-2021, Klaus Kähler Holst
 
   @brief Utility functions for Generalized Linear Models