+++ inst/include/target/pava.hpp
@@ -1,7 +1,7 @@
 /*!
   @file pava.hpp
   @author Klaus K. Holst
-  @copyright 2020, Klaus Kähler Holst
+  @copyright 2020-2021, Klaus Kähler Holst
 
   @brief Pooled adjacent violator algorithm