+++ inst/include/target/utils.hpp
@@ -1,7 +1,7 @@
 /*!
   @file utils.hpp
   @author Klaus K. Holst
-  @copyright 2020, Klaus Kähler Holst
+  @copyright 2020-2021, Klaus Kähler Holst
 
   @brief Various utility functions and constants
 
@@ -28,24 +28,29 @@ namespace target {
                   arma::vec theta);
 
   arma::umat clusterid(const arma::uvec &id);
+
   arma::mat groupsum(const arma::mat &x,
                      const arma::uvec &cluster,
                      bool reduce = true);
+
+  arma::mat interpolate(const arma::mat &input, double tau, bool locf);
+
   void fastpattern(const arma::umat &y,
                    arma::umat &pattern,
-		   arma::uvec &group,
-		   unsigned categories = 2);
-  arma::umat fastapprox(arma::vec &time,  // sorted times
+                   arma::uvec &group,
+                   unsigned categories = 2);
+
+  arma::umat fastapprox(arma::vec time,  // sorted times
                         const arma::vec &newtime,
                         bool equal = false,
-			// type: (0: nearedst, 1: right, 2: left)
+                       // type: (0: nearedst, 1: right, 2: left)
                         unsigned type = 0);
 
   double SupTest(const arma::vec &D);
   double L2Test(const arma::vec &D,
-		const arma::vec &t);
+                const arma::vec &t);
   double CramerVonMises(const arma::vec &x,
-			const arma::vec &G);
+                        const arma::vec &G);
 
   extern arma::mat const EmptyMat;
   extern arma::vec const EmptyVec;