+++ src/utils_interface.h
@@ -1,5 +1,6 @@
 #pragma once
 
+
 template <typename T>
 void Print(const std::vector<T> &x, std::string header="", std::string footer="\n") {
   Rcpp::Rcout << header;