+++ tests/testthat/test-riskreg.R
@@ -8,11 +8,10 @@ lava::distribution(m,~a) <- lava::binomial.lvm("logit")
 m <- lava::binomial.rr(m, "y","a","linpred.target","linpred.nuisance")
 d <- lava::sim(m,1e3)
 
+
 testthat::test_that("MLE convergence", {
     fit <- targeted::riskreg(y ~ a | 1 | x+z | 1, data=d, type="rr")
-    influ <- lava::iid(fit)    
+    influ <- lava::iid(fit)
     val <- sum(colMeans(influ)^2)
     testthat::expect_true(val<1e-8)
 })
-
-